Update conda environment
Explanations:

Update from Python 3.6 to 3.8. This will also update most
of the dependencies since conda packages are built for
specific minor releases. Stopped at 3.8 because
bgen 1.2.10 can't be installed on macOS with Python 3.9

numpy 1.20.0 was a huge release from Jan 2021. It
dropped support for python 3.6. Now that PR #97
has been merged and Python is being updated, it makes
sense to make the transition

https://numpy.org/devdocs/release/1.20.0-notes.html
https://github.com/numpy/numpy/releases/tag/v1.20.0

pyarrow 2.0+ requires aws-sdk-cpp to be pinned to
1.8.186. This has been backported as far back as the 3.0
release of pyarrow, so this is the new minimum

https://github.com/conda-forge/arrow-cpp-feedstock/issues/567
https://github.com/conda-forge/arrow-cpp-feedstock/pull/640

The defaults channel isn't needed to install Python and R
packages on Linux and macOS. conda-forge has everything you
need. Best case scenario the solver takes longer since it
has to update the index for the defaults channel. Worst case
(albeit unlikely) is that a defaults package will sneak in
and bork the environment (defaults and conda-forge use different
compilers and are generally incompatible)