Install and test FINEMAP
This is a simple, non-commercial test of basic functionality.
The license is noted in the CI pipeline code, and the full
license is printed to the log file. Once the CI job finishes,
the ephemeral cloud environment is deleted

http://christianbenner.com/license_finemap_v1.4.html

"""
The permission of the University is not required if the said
transmission or transference is done without financial return, the
conditions of this Licence are imposed upon the receiver of the
Software, and all original and amended source code is included in any
transmitted product.
"""